api_docs/chat_apps/chatgpt/requirements.txt
flask 
flask_cors 
werkzeug 
openai 
azure-core
azure-identity 
python-docx 
pandas
Pillow 
pytesseract 
pymupdf  
python-dotenv